refactor: CarDto 클래스를 데이터 저장 객체로 변경한다
변경 내용 :
CarDto 클래스의 이름을 GameResultOfCar 로 변경한다.
클래스 변수로 gameRound를 추가한다.
equals, hashCode, toString 메소드를 추가한다.
